Require program account to be writable in extend program data ix
