+++ programs/bpf_loader/src/lib.rs
@@ -43,8 +43,8 @@ use {
         feature_set::{
             cap_accounts_data_allocations_per_transaction, cap_bpf_program_instruction_accounts,
             disable_deploy_of_alloc_free_syscall, disable_deprecated_loader,
-            enable_bpf_loader_extend_program_data_ix,
-            error_on_syscall_bpf_function_hash_collisions, reject_callx_r10,
+            enable_bpf_loader_extend_program_ix, error_on_syscall_bpf_function_hash_collisions,
+            reject_callx_r10,
         },
         instruction::{AccountMeta, InstructionError},
         loader_instruction::LoaderInstruction,
@@ -1032,10 +1032,10 @@ fn process_loader_upgradeable_instruction(
                 }
             }
         }
-        UpgradeableLoaderInstruction::ExtendProgramData { additional_bytes } => {
+        UpgradeableLoaderInstruction::ExtendProgram { additional_bytes } => {
             if !invoke_context
                 .feature_set
-                .is_active(&enable_bpf_loader_extend_program_data_ix::ID)
+                .is_active(&enable_bpf_loader_extend_program_ix::ID)
             {
                 return Err(InstructionError::InvalidInstructionData);
             }
@@ -1046,10 +1046,11 @@ fn process_loader_upgradeable_instruction(
             }
 
             const PROGRAM_DATA_ACCOUNT_INDEX: IndexOfAccount = 0;
+            const PROGRAM_ACCOUNT_INDEX: IndexOfAccount = 1;
             #[allow(dead_code)]
             // System program is only required when a CPI is performed
-            const OPTIONAL_SYSTEM_PROGRAM_ACCOUNT_INDEX: IndexOfAccount = 1;
-            const OPTIONAL_PAYER_ACCOUNT_INDEX: IndexOfAccount = 2;
+            const OPTIONAL_SYSTEM_PROGRAM_ACCOUNT_INDEX: IndexOfAccount = 2;
+            const OPTIONAL_PAYER_ACCOUNT_INDEX: IndexOfAccount = 3;
 
             let programdata_account = instruction_context
                 .try_borrow_instruction_account(transaction_context, PROGRAM_DATA_ACCOUNT_INDEX)?;
@@ -1062,6 +1063,36 @@ fn process_loader_upgradeable_instruction(
             if !programdata_account.is_writable() {
                 ic_logger_msg!(log_collector, "ProgramData is not writable");
                 return Err(InstructionError::InvalidArgument);
+            }
+
+            let program_account = instruction_context
+                .try_borrow_instruction_account(transaction_context, PROGRAM_ACCOUNT_INDEX)?;
+            if !program_account.is_writable() {
+                ic_logger_msg!(log_collector, "Program account is not writable");
+                return Err(InstructionError::InvalidArgument);
+            }
+            if program_account.get_owner() != program_id {
+                ic_logger_msg!(log_collector, "Program account not owned by loader");
+                return Err(InstructionError::InvalidAccountOwner);
+            }
+            match program_account.get_state()? {
+                UpgradeableLoaderState::Program {
+                    programdata_address,
+                } => {
+                    if programdata_address != programdata_key {
+                        ic_logger_msg!(
+                            log_collector,
+                            "Program account does not match ProgramData account"
+                        );
+                        return Err(InstructionError::InvalidArgument);
+                    }
+
+                    drop(program_account);
+                }
+                _ => {
+                    ic_logger_msg!(log_collector, "Invalid Program account");
+                    return Err(InstructionError::InvalidAccountData);
+                }
             }
 
             let old_len = programdata_account.get_data().len();
