+++ sdk/program/src/loader_upgradeable_instruction.rs
@@ -128,7 +128,7 @@ pub enum UpgradeableLoaderInstruction {
     ///      is a ProgramData account.
     Close,
 
-    /// Extend a ProgramData account by the specified number of bytes.
+    /// Extend a program's ProgramData account by the specified number of bytes.
     /// Only upgradeable program's can be extended.
     ///
     /// The payer account must contain sufficient lamports to fund the
@@ -138,11 +138,12 @@ pub enum UpgradeableLoaderInstruction {
     ///
     /// # Account references
     ///   0. `[writable]` The ProgramData account.
-    ///   1. `[]` System program (`solana_sdk::system_program::id()`), optional, used to transfer
+    ///   1. `[writable]` The ProgramData account's associated Program account.
+    ///   2. `[]` System program (`solana_sdk::system_program::id()`), optional, used to transfer
     ///      lamports from the payer to the ProgramData account.
-    ///   2. `[signer]` The payer account, optional, that will pay necessary rent exemption costs
+    ///   3. `[signer]` The payer account, optional, that will pay necessary rent exemption costs
     ///      for the increased storage size.
-    ExtendProgramData {
+    ExtendProgram {
         /// Number of bytes to extend the program data.
         additional_bytes: u32,
     },
