+++ transaction-status/src/parse_bpf_loader.rs
@@ -155,16 +155,21 @@ pub fn parse_bpf_upgradeable_loader(
                 }),
             })
         }
-        UpgradeableLoaderInstruction::ExtendProgramData { additional_bytes } => {
+        UpgradeableLoaderInstruction::ExtendProgram { additional_bytes } => {
             check_num_bpf_upgradeable_loader_accounts(&instruction.accounts, 2)?;
             Ok(ParsedInstructionEnum {
-                instruction_type: "extendProgramData".to_string(),
+                instruction_type: "extendProgram".to_string(),
                 info: json!({
                     "additionalBytes": additional_bytes,
                     "programDataAccount": account_keys[instruction.accounts[0] as usize].to_string(),
-                    "systemProgram": account_keys[instruction.accounts[1] as usize].to_string(),
-                    "payerAccount": if instruction.accounts.len() > 2 {
+                    "programAccount": account_keys[instruction.accounts[1] as usize].to_string(),
+                    "systemProgram": if instruction.accounts.len() > 3 {
                         Some(account_keys[instruction.accounts[2] as usize].to_string())
+                    } else {
+                        None
+                    },
+                    "payerAccount": if instruction.accounts.len() > 4 {
+                        Some(account_keys[instruction.accounts[3] as usize].to_string())
                     } else {
                         None
                     },
